requirements.txt
openai
python-dotenv
termcolor
